refactor: 共通コンポーネント移動
Booksコンポーネントはプロフィールページでしか使用しないため
フォルダを作成したそこに移動させる
like出来る要素は機能追加により増える可能性があるため、likedフォルダに
格納していく
